requirements.txt
Flask
mysql-connector-python  # 如果用了MySQL